fix impl, cleanup
