+++ src/common.h
@@ -27,6 +27,8 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 #define ALIGN(data, alignto) ((data) & ~((alignto)-1))
 
+#define cpuid(level, regs) do{ __cpuid((level), (regs)[0],(regs)[1],(regs)[2],(regs)[3]); }while(0)
+
 /* Tries to find file recursively in path */
 bool find_file(std::string path, std::string file, std::string& out);
