+++ src/pci.cpp
@@ -29,7 +29,7 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include <string>
 #include <iostream>
 
-static void _pci_get_info(pci_dev*, std::string, char*);
+static void _pci_get_info(pci_dev*, std::string, std::string&);
 
 static void _init_sysfs_pci(pci_access* a){
     if(a != NULL){
@@ -43,11 +43,11 @@ static int32_t _pci_prep_rw(pci_dev* d, int32_t type){
     pci_access* a = d->access;
 
     if(a->fd == INT32_MAX || a->fd < 0){
-        char buf[1024];
+        std::string str;
 
-        _pci_get_info(d, "config", buf);
+        _pci_get_info(d, "config", str);
 
-        a->fd = open(buf, type == 0 ? O_RDONLY : O_RDWR);
+        a->fd = open(str.c_str(), type == 0 ? O_RDONLY : O_RDWR);
     }
  
     return a->fd;
@@ -55,18 +55,14 @@ static int32_t _pci_prep_rw(pci_dev* d, int32_t type){
 
 static void _read_sysfs_pci(pci_dev* d, int32_t pos, char* buf, size_t len){
     int32_t fd = _pci_prep_rw(d, 0);
-    int32_t res = pread(fd, buf, len, pos);
     
-    std::cout << " Read : " << (void*)buf << std::endl;
+    pread(fd, buf, len, pos);
 }
 
 static size_t _write_sysfs_pci(pci_dev* d, int32_t pos, char* buf, size_t len){
     int32_t fd = _pci_prep_rw(d, 1);
-    int32_t res = pwrite(fd, buf, len, pos);
 
-    std::cout << " Wrote : " << (void*)buf << std::endl;
-
-    return res; 
+    return pwrite(fd, buf, len, pos); 
 }
 
 pci_procs sysfs_procs = {
@@ -117,13 +113,13 @@ void pci_write_byte(pci_dev* dev, int32_t pos, uint8_t data){
 }
 
 void pci_write_short(pci_dev* dev, int32_t pos, uint16_t data){
-    uint16_t vdata = check_endianness() == 1 ? data : swap16(data);
+    uint16_t vdata = check_endianness() == 0 ? data : swap16(data);
     
     _pci_write(dev, &vdata, pos, sizeof(uint16_t));
 }
 
 void pci_write_long(pci_dev* dev, int32_t pos, uint32_t data){
-    uint32_t vdata = check_endianness() == 1 ? data : swap32(data);
+    uint32_t vdata = check_endianness() == 0 ? data : swap32(data);
     
     _pci_write(dev, &vdata, pos, sizeof(uint32_t));
 }
@@ -131,15 +127,7 @@ void pci_write_long(pci_dev* dev, int32_t pos, uint32_t data){
 pci_access* pci_access_alloc(){
     pci_access* a = (pci_access*)(char*)calloc(1, sizeof(pci_access));
 
-    if(a == NULL){
-        return NULL; 
-    }
-
-    /* TODO : other distros might not use misc path? */
-    a->name = "/usr/share/misc/pci.ids";
-    a->id = 0; 
-
-    return a;
+    return a != nullptr ? a : nullptr;
 }
 
 static void _pci_alloc_dev(pci_dev** d, pci_access* a){
@@ -149,15 +137,19 @@ static void _pci_alloc_dev(pci_dev** d, pci_access* a){
     dev->procs = a->procs;
 }
 
-void _pci_get_info(pci_dev* dev, std::string info, char* buf){
-    snprintf(buf, 1024, "%s/devices/%04x:%02x:%02x.%d/%s", dev->access->path.c_str(), dev->dom, dev->bus, dev->dev, dev->fun, info.c_str());
+void _pci_get_info(pci_dev* dev, std::string info, std::string& str){
+    char buf[0x400];
+
+    snprintf(buf, sizeof(buf), "%s/devices/%04x:%02x:%02x.%d/%s", dev->access->path.c_str(), dev->dom, dev->bus, dev->dev, dev->fun, info.c_str());
+
+    str = buf;
 } 
 
 pci_dev* pci_get_dev(pci_access* a, int32_t domain, int32_t dev, int32_t bus, int32_t func){
     pci_dev* d = (pci_dev*)(char*)calloc(1, sizeof(pci_dev));
 
-    if(d == NULL){
-        return NULL;
+    if(d == nullptr){
+        return nullptr;
     }
 
     _pci_alloc_dev(&d, a);
@@ -166,7 +158,7 @@ pci_dev* pci_get_dev(pci_access* a, int32_t domain, int32_t dev, int32_t bus, in
 }
 
 void pci_init_dev(pci_access* a){
-    if(a != NULL){
+    if(a != nullptr){
         a->procs = &sysfs_procs;
 
         a->procs->init(a);
