+++ src/slimbook.cpp
@@ -21,7 +21,10 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include "slimbook.h"
 #include "configuration.h"
 #include "common.h"
+#include "amdsmu.h"
+#include "pci.h"
 
+#include <cpuid.h>
 #include <sys/sysinfo.h>
 
 #include <string>
@@ -512,6 +515,107 @@ uint64_t slb_info_available_memory()
     
     return info.freeram;
 
+}
+
+/* Gets TDP from Zone 0 in CPU */
+slb_tdp_info_t _get_TDP_intel()
+{
+    #define INTEL_RAPL_PATH "/sys/class/powercap/intel-rapl/intel-rapl:0/"
+    slb_tdp_info_t tdp = {0};
+
+    if(filesystem::exists(INTEL_RAPL_PATH)){
+        string svalue;
+        read_device(INTEL_RAPL_PATH"constraint_0_power_limit_uw", svalue);
+
+        tdp.max_tdp = atoll(svalue.c_str()) / 1000000;
+    }
+    
+    return tdp;
+};
+
+/* Gets TDP from smu driver in PCI */
+slb_tdp_info_t _get_TDP_amd()
+{
+    slb_tdp_info_t tdp = {0,0,0, .type = 1};
+
+    uint32_t cpuregs[4];
+    uint32_t smuargs[2] = {0};
+
+    uint32_t family;
+    uint32_t model;
+    uint32_t design = 0;
+
+    uintptr_t addr = -1;
+
+    smu_amd* smu = NULL;
+
+    void** phys_addr = get_phys_map();
+
+    cpuid(1, cpuregs);
+
+    /* follows CPUID AMD spec standard, however, AMD CPUs after 2003 all have base family as 0xF */
+    family = ((cpuregs[0] & 0xF00) >> 8) == 0xF ? ((cpuregs[0] & 0xF00) >> 0x8) + (( cpuregs[0] & 0xFF00000) >> 0x14) : (cpuregs[0] & 0xF00) >> 8;
+    model = ((cpuregs[0] & 0xF00) >> 8) == 0xF ? ((cpuregs[0] & 0xF0000) >> 0xC ) | ((cpuregs[0] & 0xF0) >> 4): (cpuregs[0] & 0xF0) >> 4;
+
+    _get_design_amd(family, model, &design);
+
+    if(_request_addr(design, &addr, &smu, smuargs) == (uint32_t)-1){
+        return tdp;
+    }
+
+    if(addr != (uint64_t)-1){
+        _map_dev_addr(addr); 
+        _refresh_table(design, &smu, smuargs);
+
+        #define get_prop_from_offs(addr, offs) ((uint8_t)(*(float*)((uintptr_t)*(addr) + (offs))))
+
+        tdp.medium_tdp = get_prop_from_offs(phys_addr, 0x0);
+        tdp.max_tdp = get_prop_from_offs(phys_addr, 0x8);
+        tdp.low_tdp = get_prop_from_offs(phys_addr, 0x10);
+    }
+
+    pci_cleanup(smu->dev);
+    _clear_smu_amd(smu);
+    _free_map_dev();
+
+    return tdp;
+}
+
+static string _get_cpu_name(){
+    slb_smbios_entry_t* entries = nullptr;
+    int count = 0;   
+    
+    if (slb_smbios_get(&entries,&count) == 0) {
+        for (int n=0;n<count;n++) {
+            if (entries[n].type == 4) {
+                return entries[n].data.processor.version;
+            }
+        }
+    }
+
+    return "";
+}
+
+slb_tdp_info_t slb_info_get_tdp_info()
+{
+    slb_tdp_info_t tdp = {0, .type = 3};
+    int32_t cpu_type;
+    string name = _get_cpu_name();
+
+    cpu_type = name.find("AMD") != std::string::npos ? 1 : name.find("Intel") != std::string::npos ? 0 : -1;
+
+    switch(cpu_type){
+        case 0:
+            tdp = _get_TDP_intel();
+            break;
+        case 1:
+            tdp = _get_TDP_amd();
+            break;
+        default:
+            break;
+    }
+
+    return tdp;
 }
 
 const char* slb_info_keyboard_device()
