+++ src/slimbookctl.cpp
@@ -20,20 +20,10 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 #include "slimbook.h"
 #include "common.h"
+#include "amdsmu.h"
 
-#define LEGACY_PCI
-
-#include <cpuid.h>
-#ifdef LEGACY_PCI
-extern "C"{
-#include <pci/pci.h>
-}
-#else 
 #include "pci.h"
-#endif
-#include <sys/mman.h>
 #include <sys/stat.h>
-#include <fcntl.h>
 #include <sys/statvfs.h>
 #include <sys/types.h>
 #include <sys/wait.h>
@@ -43,7 +33,6 @@ extern "C"{
 #include <iostream>
 #include <iomanip>
 #include <string>
-#include <map>
 #include <vector>
 #include <fstream>
 #include <sstream>
@@ -193,422 +182,6 @@ void show_help()
     cout<<"help: show this help"<<endl;
 }
 
-/* Gets TDP from Zone 0 in CPU */
-uint64_t _get_TDP_intel(){
-    #define INTEL_RAPL_PATH "/sys/class/powercap/intel-rapl/intel-rapl:0/"
-    uint64_t uw = 0;
-
-    if(filesystem::exists(INTEL_RAPL_PATH)){
-        string svalue;
-        read_device(INTEL_RAPL_PATH"constraint_0_power_limit_uw", svalue);
-
-        uw = atoll(svalue.c_str());
-    }
-    
-    return uw;
-};
-
-typedef enum {
-    DESIGN_UNKNOWN = -1,
-    /* Family 0x17 */
-    DESIGN_RAVEN,
-    DESIGN_PICASSO,
-    DESIGN_DALI,
-    DESIGN_STARSHIP,
-    DESIGN_RENOIR,
-    DESIGN_LUCIENNE,
-    DESIGN_MATISSE,
-    DESIGN_VAN_GOGH,
-    DESIGN_MERO,
-    DESIGN_MENDOCINO,
-    /* Family 0x19 */
-    DESIGN_MILAN,
-    DESIGN_CHAGALL,
-    DESIGN_VERMEER,
-    DESIGN_BADAMI,
-    DESIGN_REMBRANDT,
-    DESIGN_CEZANNE,
-    DESIGN_STORM_PEAK,
-    DESIGN_RAPHAEL,
-    DESIGN_PHOENIX,
-    DESIGN_PHOENIX_2,
-    DESIGN_GENOA,
-    /* Family 0x1A */
-    DESIGN_TURIN,
-    DESIGN_TURIN_DENSE,
-    DESIGN_STRIX_POINT_1,
-    DESIGN_STRIX_POINT_2,
-    DESIGN_STRIX_HALO,
-    DESIGN_GRANITE_RIDGE,
-    DESIGN_FIRE_RANGE,
-    DESIGN_KRACKAN_POINT_1,
-    DESIGN_SARLAK,
-}_amd_design;
-
-
-/* Uses switch range C++ extension, not standard but impl in important compilers */
-static void _get_design_amd(uint32_t family, uint32_t model, uint32_t* design){
-    switch(family){
-        case 0x17:
-            switch(model){
-                case 0x11:
-                    *design = DESIGN_RAVEN;
-                    return;
-                case 0x18:
-                    *design = DESIGN_PICASSO;
-                    return;
-                case 0x20:
-                    *design = DESIGN_DALI;
-                    return;
-                case 0x31:
-                    *design = DESIGN_STARSHIP;
-                    return;
-                case 0x47:
-                case 0x87:
-                    // figure out
-                    return;
-                case 0x60:
-                    *design = DESIGN_RENOIR;
-                    return;
-                case 0x68:
-                    *design = DESIGN_LUCIENNE;
-                    return;
-                case 0x71:
-                    *design = DESIGN_MATISSE;
-                    return;
-                case 0x90:
-                    *design = DESIGN_VAN_GOGH;
-                    return;
-                case 0x98:
-                    *design = DESIGN_MERO;
-                case 0xA0:
-                    *design = DESIGN_MENDOCINO;
-                    return;
-                default:
-                    *design = DESIGN_UNKNOWN;
-                    return;
-            }
-            return;
-        case 0x19:
-            switch(model){
-                case 0x1:
-                    *design = DESIGN_MILAN;
-                    return;
-                case 0x8:
-                    *design = DESIGN_CHAGALL;
-                    return;
-                case 0x20 ... 0x2F :
-                    *design = DESIGN_VERMEER;
-                    return;
-                case 0x30 ... 0x3F:
-                    *design = DESIGN_BADAMI;
-                    return;
-                case 0x40 ... 0x4F:
-                    *design = DESIGN_REMBRANDT;
-                    return;
-                case 0x50 ... 0x5F:
-                    *design = DESIGN_CEZANNE;
-                    return;
-                case 0x10 ... 0x1F:
-                    *design = DESIGN_STORM_PEAK;
-                    return;
-                case 0x60 ... 0x6F:
-                    *design = DESIGN_RAPHAEL;
-                    return;
-                case 0x70 ... 0x77:
-                    *design = DESIGN_PHOENIX;
-                    return;
-                case 0x78 ... 0x7F:
-                    *design = DESIGN_PHOENIX_2;
-                    return;
-                case 0xA0 ... 0xAF:
-                    *design = DESIGN_GENOA;
-                    return;
-                default:
-                    *design = DESIGN_UNKNOWN;
-                    return;
-            } 
-            case 0x1A:
-                switch(model){
-                    case 0x0 ... 0xF:
-                        *design = DESIGN_TURIN;
-                        return;
-                    case 0x10 ... 0x1F:
-                        *design = DESIGN_TURIN_DENSE;
-                        return;
-                    case 0x20 ... 0x2F:
-                        *design = DESIGN_STRIX_POINT_1;
-                        return;
-                    case 0x30 ... 0x37:
-                        *design = DESIGN_STRIX_POINT_2;
-                        return;
-                    case 0x38 ... 0x3F:
-                        *design = DESIGN_STRIX_HALO;
-                        return;
-                    case 0x40 ... 0x4F:
-                        *design = DESIGN_GRANITE_RIDGE;
-                        return;
-                    case 0x50 ... 0x5F:
-                        *design = DESIGN_FIRE_RANGE;
-                        return;
-                    case 0x60 ... 0x6F:
-                        *design = DESIGN_KRACKAN_POINT_1;
-                        return;
-                    case 0x70 ... 0x77:
-                        *design = DESIGN_SARLAK;
-                        return;
-                    default:
-                        *design = DESIGN_UNKNOWN;
-                        return;        
-                }  
-
-            /* TODO: Support older families than Zen? */
-            default:
-                *design = DESIGN_UNKNOWN;
-                return;
-    }   
-}
-
-#define cpuid(level, regs) do{ __cpuid((level), (regs)[0],(regs)[1],(regs)[2],(regs)[3]); }while(0)
-
-typedef struct _smu_amd{
-    struct pci_dev* dev;
-    uint32_t msg;
-    uint32_t res;
-    uint32_t arg_base;
-}smu_amd;
-
-#define MSG_MSG_ADDR 0x03B10A20
-#define MSG_RES_ADDR 0x03B10A80
-#define MSG_ARG_BASE_ADDR 0x03B10A88
-
-smu_amd* _get_smu_amd(){
-    #ifndef LEGACY_PCI
-    pci_access* o = pci_access_alloc(); //pci_access_alloc
-    pci_dev* dev;
-    #else
-    struct pci_access* o = pci_alloc();
-    struct pci_dev* dev;
-    #endif
-    smu_amd* smu;
-    
-    smu = (smu_amd*)((char*)malloc(sizeof(smu_amd)));
-
-    #ifndef LEGACY_PCI
-    pci_init_dev(o); //pci_init_dev
-    #else
-    pci_init(o);
-    #endif
-
-    dev = pci_get_dev(o, 0,0,0,0); //pci_get_dev
-
-    smu->dev = dev;
-    smu->msg = MSG_MSG_ADDR;
-    smu->res = MSG_RES_ADDR;
-    smu->arg_base = MSG_ARG_BASE_ADDR;
-    
-    return smu;
-}
-
-void _clear_smu_amd(smu_amd* smu){
-    if(smu != nullptr){
-        free(smu);
-    }
-}
-
-#define DEV_PCI_REG_ADDR_ADDR 0xB8
-#define DEV_PCI_REG_DATA_ADDR 0xBC
-
-static uint32_t _pci_reg_rd(pci_dev* dev, uint32_t addr){
-    pci_write_long(dev, DEV_PCI_REG_ADDR_ADDR, ALIGN(addr, 4));
-    return pci_read_long(dev, DEV_PCI_REG_DATA_ADDR);
-}
-
-static void _pci_reg_wr(pci_dev* dev, uint32_t addr, uint32_t data){
-    pci_write_long(dev, DEV_PCI_REG_ADDR_ADDR, addr);
-    pci_write_long(dev, DEV_PCI_REG_DATA_ADDR, data);
-}
-
-#define MSG_ARG_ADDR(base, num) ((base) + 4 * (num))
-
-uint32_t _smu_amd_send_req(smu_amd* smu, uint32_t msg, uint32_t* args){
-    uint32_t res = 0;
-
-    _pci_reg_wr(smu->dev, smu->res, 0);
-
-    _pci_reg_wr(smu->dev, MSG_ARG_ADDR(smu->arg_base, 0), args[0]);
-    _pci_reg_wr(smu->dev, MSG_ARG_ADDR(smu->arg_base, 1), args[1]);
-
-    // cout << args[0] << endl;
-    // cout << args[1] << endl;
-
-    _pci_reg_wr(smu->dev, smu->msg, msg);
-
-    while(res == 0){
-        res = _pci_reg_rd(smu->dev, smu->res);
-    }
-
-    args[0] = _pci_reg_rd(smu->dev, MSG_ARG_ADDR(smu->arg_base, 0));
-    args[1] = _pci_reg_rd(smu->dev, MSG_ARG_ADDR(smu->arg_base, 1));
-
-    //    cout << std::hex << args[0] << endl;
-    //    cout << std::hex <<  args[1] << endl;
-
-    return res;
-}
-
-static void* phys_map = MAP_FAILED;
-
-void _map_dev_addr(uintptr_t addr){
-    int dev_fd = open("/dev/mem", O_RDONLY);
-
-    if(dev_fd > 0){
-        phys_map = mmap(NULL, 4096, PROT_READ, MAP_SHARED, dev_fd, addr);
-        close(dev_fd);
-    }
-}
-
-void _free_map_dev(){
-    if(phys_map != MAP_FAILED){
-        munmap((void*)phys_map, 4096);
-    }
-}
-
-uint32_t _request_addr(uint32_t design, uintptr_t* addr, smu_amd** smu, uint32_t* smuargs){
-    uint32_t table_msg = -1;
-
-    switch(design){
-        case DESIGN_RAVEN:
-        case DESIGN_PICASSO:
-        case DESIGN_DALI:
-            smuargs[0] = 3;
-            table_msg = 0xB;
-            break;
-        
-        case DESIGN_RENOIR:
-        case DESIGN_LUCIENNE:
-        case DESIGN_CEZANNE:
-        case DESIGN_REMBRANDT:
-        case DESIGN_PHOENIX:
-        case DESIGN_PHOENIX_2:
-            table_msg = 0x66;
-            break;
-        default:
-            return -1;
-            break;
-    }
-
-    if(table_msg != (uint32_t)-1){
-        uint32_t res;
-
-        *smu = _get_smu_amd();
-        res = _smu_amd_send_req(*smu, table_msg, smuargs);
-        
-        if(res != 1){
-            return -1;
-        }
-
-        switch(design){
-            case DESIGN_REMBRANDT:
-            case DESIGN_PHOENIX:
-            case DESIGN_PHOENIX_2:
-                *addr = (uint64_t) smuargs[1] << 32 | smuargs[0];
-                return 0;
-            default:
-                *addr = (uint64_t)smuargs[0];
-                return 0;
-        }
-    }
-    else{
-        return -1;
-    }
-
-}
-
-uint32_t _refresh_table(uint32_t design, smu_amd** smu, uint32_t* smuargs){
-    uint32_t table_msg = -1;
-
-    switch(design){
-        case DESIGN_RAVEN:
-        case DESIGN_PICASSO:
-        case DESIGN_DALI:
-            smuargs[0] = 3;
-            table_msg = 0x3D;
-            break;
-        
-        case DESIGN_RENOIR:
-        case DESIGN_LUCIENNE:
-        case DESIGN_CEZANNE:
-        case DESIGN_REMBRANDT:
-        case DESIGN_PHOENIX:
-        case DESIGN_PHOENIX_2:
-            table_msg = 0x65;
-            break;
-        default:
-            return -1;
-            break;
-    }
-
-    if(table_msg != (uint32_t)-1){
-        uint32_t res;
-
-        res = _smu_amd_send_req(*smu, table_msg, smuargs);
-
-        if(res == 253){
-            usleep(200 * 1000);
-            res = _smu_amd_send_req(*smu, table_msg, smuargs);
-        }
-
-        if(res == 253){
-            return -1;
-        }
-    }
-
-    return 0;
-}
-
-float _get_TDP_amd(){
-    float uw;
-
-    uint32_t cpuregs[4];
-    uint32_t smuargs[2] = {0};
-
-    uint32_t family;
-    uint32_t model;
-    uint32_t design = 0;
-
-    uintptr_t addr = -1;
-
-    smu_amd* smu = NULL;
-
-    cpuid(1, cpuregs);
-
-    /* follows CPUID AMD spec standard, however, AMD CPUs after 2003 all have base family as 0xF */
-    family = ((cpuregs[0] & 0xF00) >> 8) == 0xF ? ((cpuregs[0] & 0xF00) >> 0x8) + (( cpuregs[0] & 0xFF00000) >> 0x14) : (cpuregs[0] & 0xF00) >> 8;
-    model = ((cpuregs[0] & 0xF00) >> 8) == 0xF ? ((cpuregs[0] & 0xF0000) >> 0xC ) | ((cpuregs[0] & 0xF0) >> 4): (cpuregs[0] & 0xF0) >> 4;
-
-    _get_design_amd(family, model, &design);
-
-    if(_request_addr(design, &addr, &smu, smuargs) == (uint32_t)-1){
-        return 0.0f;
-    }
-
-    if(addr != (uint64_t)-1){
-        _map_dev_addr(addr); 
-        _refresh_table(design, &smu, smuargs);
-
-        uw = *(float*)((uintptr_t)phys_map + 0x8);
-    }
-
-    #ifndef LEGACY_PCI
-    pci_cleanup(smu->dev);
-    #endif
-    _clear_smu_amd(smu);
-    _free_map_dev();
-
-    return uw;
-}
-
 string get_info()
 {
     stringstream sout;
@@ -713,33 +286,38 @@ string get_info()
     
     slb_smbios_entry_t* entries = nullptr;
     int count = 0;
-    int cpu_type = 1;
-
 
     if (slb_smbios_get(&entries,&count) == 0) {
         for (int n=0;n<count;n++) {
             if (entries[n].type == 4) {
                 string name = trim(entries[n].data.processor.version);
                 int count = entries[n].data.processor.threads;
-                uint32_t wUsage = 0;
+                slb_tdp_info_t tdp = {0};
                  
                 sout<<"cpu:"<<name<<" x "<<count<<endl;
 
-                switch(cpu_type){
+                tdp = slb_info_get_tdp_info();
+
+                sout << "TDP: ";
+
+                switch(tdp.type){
                     case 0:
-                        wUsage = _get_TDP_intel() / 1000000;
+                        sout << tdp.max_tdp << " W" << endl;
                         break;
+
                     case 1:
-                        wUsage = _get_TDP_amd();
+                        sout << "\nLOW: " << (int)tdp.low_tdp << " W";
+                        sout << "\nMEDIUM: " << (int)tdp.medium_tdp << " W";
+                        sout << "\nMAX: " << (int)tdp.max_tdp << " W";
+
+                        sout << "\n";
+
                         break;
+
                     default:
                         break;
                 }
-
-                sout << "TDP/PPT: " << wUsage << "W" << endl;
-
-
-            }
+            }  
             
             if (entries[n].type == 17) {
                 if (entries[n].data.memory_device.type > 2) {
